Allow 9 levels of inheritance when linting
Testing for future versions of Python exposed that the standard library changes
the inheritance underneath collections.MutableSet, resulting in failed tests
(3.6a and 3.7a).

Ruruki only implements an interface followed by a concrete implementation on
top of the standard library, this should not be a difficult amount of
complexity to take in for a developer.